extended  - pos

+++ UTrans/network_architecture/nnformergt_ext.py
@@ -383,7 +383,7 @@ class SwinTransformerBlock(nn.Module):
         gt = rearrange(gt, "b (n g) c -> (b n) g c",g=ngt, c=C)
 
         # New vts
-        vts_ = vts.clone()
+        vts_ = vts.clone().half()
         print("t vts_",type(vts_))
         print("t vt",type(vt))
         vts_[vt_pos] = vt
